refactor(deno-runtime): extract main handler loop from main.ts
squash into previous PR
