+++ packages/apps/deno-runtime/lib/require.ts
@@ -7,8 +7,7 @@ export const require = (mod: string) => {
 	// However, the import maps are configured to look at the source folder for typescript files, but during
 	// runtime those files are not available
 	if (mod.startsWith('@rocket.chat/apps-engine')) {
-		// Only remove "src/" substring when it comes after "apps-engine/"
-		mod = import.meta.resolve(mod).replace('file://', '').replace('apps-engine/src/', 'apps-engine/');
+		mod = import.meta.resolve(mod).replace('file://', '');
 	}
 
 	return _require(mod);
